removing unused files and renaming new
