+++ src/common/config.json
@@ -7,17 +7,7 @@
 	"file": {
 		"category": "io",
 		"key": "$File",
-		"description": "Import of UFT-8 Text file"
-	},
-	"yamlFile": {
-		"category": "io",
-		"key": "$YamlFile",
-		"description": "Import of Yaml File"
-	},
-	"jsonFile": {
-		"category": "io",
-		"key": "$JsonFile",
-		"description": "Import of stringified JSON File"
+		"description": "Import of UFT-8 file. Support import of YAML, JSON and pure text."
 	},
 	"sum": {
 		"category": "math",
