--- src/io/logic.puml
@@ -1,32 +1,0 @@
-@startuml
-start
-repeat
-group Initiate
-	#lightgreen:Input Source;
-	if (Is source set?) then (Yes)
-	else (No)
-		:Set location of execution as source;
-	endif
-	#lightgreen:Input Relative File Path;
-	:Resolve Path;
-	:Read File;
-endgroup
-
-
-group Transform
-	if (Need transformation?) then (Yes)
-		switch (FileType is...)
-		case (JSON)
-			:JSON.parse;
-		case (YAML)
-			:YAML.load;
-		endswitch
-	else (No)
-		:Forward;
-	endif
-end group
-repeat while (Repeat) is (Yes)
-->no;
-:Return Data;
-stop
-@enduml
